sky and gradient initalization refactored to use common code.
