+++ ToolKit/Sky.cpp
@@ -10,6 +10,7 @@
 
 namespace ToolKit
 {
+
   SkyBase::SkyBase()
   {
     ParameterConstructor();
@@ -18,7 +19,34 @@ namespace ToolKit
 
   EntityType SkyBase::GetType() const { return EntityType::Entity_SkyBase; }
 
-  void SkyBase::Init() {}
+  void SkyBase::Init()
+  {
+    if (m_initialized)
+    {
+      return;
+    }
+
+    EnvironmentComponentPtr envComp = GetComponent<EnvironmentComponent>();
+    if (envComp == nullptr)
+    {
+      // Create a default environment component.
+      envComp = std::make_shared<EnvironmentComponent>();
+      
+      TextureManager* tman = GetTextureManager();
+      HdriPtr defHdr =
+          tman->Create<Hdri>(tman->GetDefaultResource(ResourceType::Hdri));
+      defHdr->Init();
+
+      envComp->SetHdriVal(defHdr);
+      AddComponent(envComp);
+    }
+
+    envComp->m_entity = this;
+    envComp->Init(false);
+
+    // Do not expose environment component
+    envComp->m_localData.ExposeByCategory(false, EnvironmentComponentCategory);
+  }
 
   void SkyBase::ReInit()
   {
@@ -67,13 +95,20 @@ namespace ToolKit
 
   void SkyBase::ParameterEventConstructor() {}
 
+  void SkyBase::ConstructSkyMaterial(ShaderPtr vertexPrg, ShaderPtr fragPrg)
+  {
+    m_skyboxMaterial                   = std::make_shared<Material>();
+
+    EnvironmentComponentPtr envComp    = GetComponent<EnvironmentComponent>();
+    m_skyboxMaterial->m_cubeMap        = envComp->GetHdriVal()->m_cubemap;
+    m_skyboxMaterial->m_vertexShader   = vertexPrg;
+    m_skyboxMaterial->m_fragmentShader = fragPrg;
+    m_skyboxMaterial->GetRenderState()->cullMode = CullingType::TwoSided;
+    m_skyboxMaterial->Init();
+  }
+
   Sky::Sky()
   {
-    EnvironmentComponent* envComp = new EnvironmentComponent;
-    envComp->SetHdriVal(GetTextureManager()->Create<Hdri>(
-        TexturePath("defaultHDRI.hdr", true)));
-    AddComponent(envComp);
-
     ParameterConstructor();
     ParameterEventConstructor();
   }
@@ -89,20 +124,7 @@ namespace ToolKit
       return;
     }
 
-    // Environment Component
-    EnvironmentComponentPtr envComp = GetComponent<EnvironmentComponent>();
-    envComp->m_entity               = this;
-    envComp->GetHdriVal()->UnInit();
-    envComp->SetExposureVal(GetExposureVal());
-    envComp->Init(true);
-
-    // Do not expose environment component
-    envComp->ParamPositionOffset().m_exposed = false;
-    envComp->ParamSize().m_exposed           = false;
-    envComp->ParamHdri().m_exposed           = false;
-    envComp->ParamIlluminate().m_exposed     = false;
-    envComp->ParamIntensity().m_exposed      = false;
-    envComp->ParamExposure().m_exposed       = false;
+    SkyBase::Init();
 
     // Skybox material
     ShaderPtr vert = GetShaderManager()->Create<Shader>(
@@ -111,14 +133,7 @@ namespace ToolKit
     ShaderPtr frag = GetShaderManager()->Create<Shader>(
         ShaderPath("skyboxFrag.shader", true));
 
-    m_skyboxMaterial                   = std::make_shared<Material>();
-    m_skyboxMaterial->m_cubeMap        = envComp->GetHdriVal()->m_cubemap;
-    m_skyboxMaterial->m_vertexShader   = vert;
-    m_skyboxMaterial->m_fragmentShader = frag;
-
-    RenderState* rs                    = m_skyboxMaterial->GetRenderState();
-    rs->cullMode                       = CullingType::TwoSided;
-    m_skyboxMaterial->Init();
+    ConstructSkyMaterial(vert, frag);
 
     m_initialized = true;
   }
@@ -136,17 +151,14 @@ namespace ToolKit
 
   void Sky::ParameterConstructor()
   {
-    Exposure_Define(GetComponent<EnvironmentComponent>()->GetExposureVal(),
+    Exposure_Define(1.0f,
                     "Sky",
                     90,
                     true,
                     true,
                     {false, true, 0.0f, 50.0f, 0.05f});
-    Hdri_Define(GetComponent<EnvironmentComponent>()->GetHdriVal(),
-                "Sky",
-                90,
-                true,
-                true);
+
+    Hdri_Define(nullptr, "Sky", 90, true, true);
 
     SetNameVal("Sky");
 
@@ -178,227 +190,6 @@ namespace ToolKit
           GetComponent<EnvironmentComponent>()->SetHdriVal(
               std::get<HdriPtr>(newVal));
         });
-  }
-
-  GradientSky::GradientSky()
-  {
-    ParameterConstructor();
-    ParameterEventConstructor();
-  }
-
-  GradientSky::~GradientSky() {}
-
-  EntityType GradientSky::GetType() const
-  {
-    return EntityType::Entity_GradientSky;
-  }
-
-  void GradientSky::Init()
-  {
-    if (m_initialized)
-    {
-      return;
-    }
-
-    // Skybox material
-    ShaderPtr vert = GetShaderManager()->Create<Shader>(
-        ShaderPath("skyboxVert.shader", true));
-
-    ShaderPtr frag = GetShaderManager()->Create<Shader>(
-        ShaderPath("gradientSkyboxFrag.shader", true));
-
-    m_skyboxMaterial                   = std::make_shared<Material>();
-    m_skyboxMaterial->m_vertexShader   = vert;
-    m_skyboxMaterial->m_fragmentShader = frag;
-
-    RenderState* rs                    = m_skyboxMaterial->GetRenderState();
-    rs->cullMode                       = CullingType::TwoSided;
-    m_skyboxMaterial->Init();
-
-    RenderTask task {[this](Renderer* renderer) -> void
-                     {
-                       // Render gradient to cubemap and store the output
-                       GenerateGradientCubemap();
-
-                       m_onInit = true;
-
-                       // Create irradiance map from cubemap and set
-                       GenerateIrradianceCubemap();
-                     }};
-
-    GetRenderSystem()->AddRenderTask(task);
-  }
-
-  MaterialPtr GradientSky::GetSkyboxMaterial()
-  {
-    Init();
-
-    m_skyboxMaterial->m_fragmentShader->SetShaderParameter(
-        "topColor",
-        ParameterVariant(GetTopColorVal()));
-
-    m_skyboxMaterial->m_fragmentShader->SetShaderParameter(
-        "middleColor",
-        ParameterVariant(GetMiddleColorVal()));
-
-    m_skyboxMaterial->m_fragmentShader->SetShaderParameter(
-        "bottomColor",
-        ParameterVariant(GetBottomColorVal()));
-
-    m_skyboxMaterial->m_fragmentShader->SetShaderParameter(
-        "exponent",
-        ParameterVariant(GetGradientExponentVal()));
-
-    return m_skyboxMaterial;
-  }
-
-  CubeMapPtr GradientSky::GetIrradianceMap() { return m_irradianceMap; }
-
-  void GradientSky::ParameterConstructor()
-  {
-    TopColor_Define(Vec3(0.3f, 0.3f, 1.0f), "Sky", 90, true, true, {true});
-    MiddleColor_Define(Vec3(1.0f, 1.0f, 0.8f), "Sky", 90, true, true, {true});
-    BottomColor_Define(Vec3(0.5f, 0.3f, 0.1f), "Sky", 90, true, true, {true});
-    GradientExponent_Define(0.3f,
-                            "Sky",
-                            90,
-                            true,
-                            true,
-                            {false, true, 0.0f, 10.0f, 0.02f});
-
-    IrradianceResolution_Define(64.0f,
-                                "Sky",
-                                90,
-                                true,
-                                true,
-                                {false, true, 32.0f, 2048.0f, 2.0f});
-
-    SetNameVal("Gradient Sky");
-  }
-
-  void GradientSky::ParameterEventConstructor() {}
-
-  void GradientSky::GenerateGradientCubemap()
-  {
-    RenderTask task = {
-        [this](Renderer* renderer) -> void
-        {
-          FramebufferPtr fb = std::make_shared<Framebuffer>();
-          fb->Init({m_size, m_size, false, true});
-
-          const RenderTargetSettigs set = {0,
-                                           GraphicTypes::TargetCubeMap,
-                                           GraphicTypes::UVClampToEdge,
-                                           GraphicTypes::UVClampToEdge,
-                                           GraphicTypes::UVClampToEdge,
-                                           GraphicTypes::SampleLinear,
-                                           GraphicTypes::SampleLinear,
-                                           GraphicTypes::FormatRGB,
-                                           GraphicTypes::FormatRGB,
-                                           GraphicTypes::TypeUnsignedByte};
-
-          RenderTargetPtr cubemap =
-              std::make_shared<RenderTarget>(m_size, m_size, set);
-
-          cubemap->Init();
-
-          // Create material
-          m_skyboxMaterial->m_fragmentShader->SetShaderParameter(
-              "topColor",
-              ParameterVariant(GetTopColorVal()));
-
-          m_skyboxMaterial->m_fragmentShader->SetShaderParameter(
-              "middleColor",
-              ParameterVariant(GetMiddleColorVal()));
-
-          m_skyboxMaterial->m_fragmentShader->SetShaderParameter(
-              "bottomColor",
-              ParameterVariant(GetBottomColorVal()));
-
-          m_skyboxMaterial->m_fragmentShader->SetShaderParameter(
-              "exponent",
-              ParameterVariant(GetGradientExponentVal()));
-
-          renderer->EnableDepthTest(false);
-
-          // Views for 6 different angles
-          CameraPtr cam = std::make_shared<Camera>();
-          cam->SetLens(glm::radians(90.0f), 1.0f, 0.1f, 10.0f);
-
-          Mat4 views[] = {
-              glm::lookAt(ZERO,
-                          Vec3(1.0f, 0.0f, 0.0f),
-                          Vec3(0.0f, -1.0f, 0.0f)),
-              glm::lookAt(ZERO,
-                          Vec3(-1.0f, 0.0f, 0.0f),
-                          Vec3(0.0f, -1.0f, 0.0f)),
-              glm::lookAt(ZERO,
-                          Vec3(0.0f, -1.0f, 0.0f),
-                          Vec3(0.0f, 0.0f, -1.0f)),
-              glm::lookAt(ZERO, Vec3(0.0f, 1.0f, 0.0f), Vec3(0.0f, 0.0f, 1.0f)),
-              glm::lookAt(ZERO,
-                          Vec3(0.0f, 0.0f, 1.0f),
-                          Vec3(0.0f, -1.0f, 0.0f)),
-              glm::lookAt(ZERO,
-                          Vec3(0.0f, 0.0f, -1.0f),
-                          Vec3(0.0f, -1.0f, 0.0f))};
-
-          for (int i = 0; i < 6; ++i)
-          {
-            Vec3 pos;
-            Quaternion rot;
-            Vec3 sca;
-            DecomposeMatrix(views[i], &pos, &rot, &sca);
-
-            cam->m_node->SetTranslation(ZERO, TransformationSpace::TS_WORLD);
-            cam->m_node->SetOrientation(rot, TransformationSpace::TS_WORLD);
-            cam->m_node->SetScale(sca);
-
-            fb->SetAttachment(Framebuffer::Attachment::ColorAttachment0,
-                              cubemap,
-                              0,
-                              -1,
-                              (Framebuffer::CubemapFace) i);
-
-            renderer->SetFramebuffer(fb, true, Vec4(0.0f));
-            renderer->DrawCube(cam.get(), m_skyboxMaterial);
-          }
-
-          renderer->EnableDepthTest(true);
-
-          // Take the ownership of render target.
-          m_skyboxMap = std::make_shared<CubeMap>(cubemap->m_textureId);
-          cubemap->m_textureId = 0;
-          cubemap              = nullptr;
-        }};
-
-    GetRenderSystem()->AddRenderTask(task);
-  }
-
-  void GradientSky::GenerateIrradianceCubemap()
-  {
-    RenderTask task = {[this](Renderer* renderer) -> void
-                       {
-                         TexturePtr irradianceMap =
-                             renderer->GenerateEnvIrradianceMap(
-                                 m_skyboxMap,
-                                 (uint) GetIrradianceResolutionVal(),
-                                 (uint) GetIrradianceResolutionVal());
-
-                         // Take the ownership of render target.
-                         m_irradianceMap = std::make_shared<CubeMap>(
-                             irradianceMap->m_textureId);
-                         irradianceMap->m_textureId = 0;
-                         irradianceMap              = nullptr;
-
-                         if (m_onInit)
-                         {
-                           m_initialized = true;
-                           m_onInit      = false;
-                         }
-                       }};
-
-    GetRenderSystem()->AddRenderTask(task);
   }
 
 } // namespace ToolKit
